レビューグラフを別ファイルに記述
