+++ src/resources/js/Pages/Farm/Detail.tsx
@@ -5,6 +5,7 @@ import { StarIcon, EditIcon } from "@chakra-ui/icons";
 import { router } from "@inertiajs/react";
 import FarmImageList from "@/Components/Organisms/FarmImageList";
 import FarmList from "@/Components/Organisms/FarmList";
+import FarmRatingGraph from "@/Components/Organisms/FarmRatingGraph";
 
 type State = { id: number; name: string };
 type FarmImages = { id: number; farm_id: number; url: string };
@@ -22,7 +23,6 @@ type Review = {
     end_date: string;
     application_method_id?: number | null;
     application_method_name?: string | null;
-    other_application_method?: string | null;
     application_method_other?: string | null;
     application_method?: { id: number; name: string } | null;
     review_user?: { id: number; nickname: string } | null;
@@ -46,75 +46,21 @@ type Farm = {
 
 type DetailProps = { farm: Farm };
 
-const RatingSummary: React.FC<{ reviews?: Review[] }> = ({ reviews }) => {
-    const ratings = reviews?.map((r) => r.farm_rating) ?? [];
-    const total = ratings.length;
-
-    const { avg, percents } = useMemo(() => {
-        if (total === 0) return { avg: 0, counts: [0, 0, 0, 0, 0], percents: [0, 0, 0, 0, 0] };
-
-        const avg = Math.round((ratings.reduce((a, b) => a + b, 0) / total) * 10) / 10;
-        const counts = [5, 4, 3, 2, 1].map((n) => ratings.filter((r) => r === n).length);
-        const max = Math.max(...counts);
-        const percents = counts.map((c) => (max ? (c / max) * 100 : 0));
-        return { avg, counts, percents };
-    }, [ratings, total]);
-
-    return (
-        <HStack align="flex-start" spacing={8} mt={4} mb={6}>
-            <VStack align="flex-start">
-                <Text fontSize="6xl" fontWeight="bold" lineHeight="1">{avg.toFixed(1)}</Text>
-                <HStack>
-                    {Array(5)
-                        .fill(0)
-                        .map((_, i) => (
-                            <StarIcon
-                                key={i}
-                                color={i < Math.round(avg) ? "green.500" : "gray.300"}
-                                boxSize={5}
-                            />
-                        ))}
-                </HStack>
-                <Text color="gray.600" fontSize="sm">{total.toLocaleString()} 件のレビュー</Text>
-            </VStack>
-
-            <VStack flex="1" spacing={2} align="stretch">
-                {[5, 4, 3, 2, 1].map((star, idx) => (
-                    <HStack key={star} spacing={3}>
-                        <Text w="16px" textAlign="right" fontSize="sm">{star}</Text>
-                        <Progress
-                            value={percents[idx]}
-                            flex="1"
-                            size="md"
-                            borderRadius="md"
-                            colorScheme="green"
-                            bg="gray.200"
-                        />
-                    </HStack>
-                ))}
-            </VStack>
-        </HStack>
-    );
-};
-
 const Detail = ({ farm }: DetailProps) => {
-    const OTHER_ID = 99;
-    const OTHER_LABEL = "その他";
+    const OtherId = 99;
+    const OtherLabel = "その他";
 
     const renderApplicationMethod = (review: Review): string => {
-        const name =
-            (review.application_method_name ?? review.application_method?.name ?? "").trim();
+        const name = (review.application_method_name ?? review.application_method?.name ?? "").trim();
 
-        const other =
-            (review.application_method_other ?? review.other_application_method ?? "").trim();
+        const other = (review.application_method_other ?? "").trim();
 
-        const isOther = review.application_method_id === OTHER_ID || name === OTHER_LABEL;
+        const isOther = review.application_method_id === OtherId || name === OtherLabel;
 
         if (isOther) {
-            return other ? `${OTHER_LABEL}：「${other}」` : OTHER_LABEL;
+            return other ? other : OtherLabel;
         }
-        if (name) return name;
-        return "未入力";
+        return name;
     };
 
     return (
@@ -149,7 +95,7 @@ const Detail = ({ farm }: DetailProps) => {
                     レビュー
                 </Heading>
 
-                <RatingSummary reviews={farm.reviews} />
+                <FarmRatingGraph reviews={farm.reviews ?? []} />
 
                 <Box display="flex" justifyContent="space-between" mb={3}>
                     {farm.reviews?.length === 0 ? "レビューの登録なし" : `${farm.reviews?.length}件`}
